double all the things

+++ src/main/java/com/contaazul/coverage/pullrequest/analyser/LineAnalyser.java
@@ -6,7 +6,7 @@ import com.contaazul.coverage.pullrequest.Cobertura;
 public class LineAnalyser {
 	// TODO side-effects!
 	public final void analyse(int line, LineCoverager coverager, Cobertura chunkCoverage) {
-		final Integer lineCoverage = coverager.getLineCoverage( line );
+		final Double lineCoverage = coverager.getLineCoverage( line );
 		if (lineCoverage != null)
 			chunkCoverage.incrementCoverage( line, lineCoverage );
 	}
